Introduce BindAddress
BindAddress is the server side equivalent of the AdvertisedAddress
and allows to specify where to bind the server to.
