+++ crates/types/src/lib.rs
@@ -25,6 +25,7 @@ pub mod invocation;
 pub mod journal;
 pub mod logs;
 pub mod message;
+pub mod net;
 pub mod nodes_config;
 pub mod partition_table;
 pub mod retries;
